requirements.txt
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
pillow==10.0.0
tensorflow==2.13.0
keras==2.13.1
flask==2.3.3
pytest==7.4.0
opencv-python==4.8.0.76 